Updates azure workflow; Removes Gemfile
Liquid throws exception when we add Gemfile to repo. On default sample, there is no Gemfile, and it works fine on local machine, but does not when add Gemfile. But in workflow we defined "bundle update" step that required, Gemfile. As workaround trying to comment "bundle update" step (no necessity for now).